requirements.txt
asgiref==3.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.11.1
django-anymail==13.0.1
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.3.1
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.8.2
